# lib/api-zod/src/generated/types/createExperienceRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateExperienceRequestType } from "./createExperienceRequestType";

export interface CreateExperienceRequest {
  /**
   * @minLength 1
   * @maxLength 200
   */
  title: string;
  /**
   * @minLength 1
   * @maxLength 2000
   */
  description: string;
  type: CreateExperienceRequestType;
  /** @minLength 1 */
  category: string;
  /** @minLength 1 */
  date: string;
  /**
   * @minLength 1
   * @maxLength 100
   */
  city: string;
  /**
   * @minLength 1
   * @maxLength 100
   */
  country: string;
  /** @minimum 2 */
  max_participants?: number | null;
  /**
   * @minimum 10
   * @maximum 500
   */
  xp_reward?: number;
}