# lib/api-zod/src/generated/types/experienceCard.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ExperienceCardType } from "./experienceCardType";

export interface ExperienceCard {
  id: number;
  title: string;
  type: ExperienceCardType;
  category: string;
  date: string;
  city: string;
  xp_reward: number;
  participant_count: number;
  max_participants?: number | null;
  creator_name: string;
  creator_username: string;
  status: string;
}