# lib/api-zod/src/generated/types/experienceDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ExperienceCreator } from "./experienceCreator";
import type { ExperienceDetailType } from "./experienceDetailType";

export interface ExperienceDetail {
  id: number;
  title: string;
  description: string;
  type: ExperienceDetailType;
  category: string;
  date: string;
  city: string;
  country: string;
  xp_reward: number;
  max_participants?: number | null;
  participant_count: number;
  status: string;
  creator: ExperienceCreator;
  joined: boolean;
  participation_status?: string | null;
  created_at: string;
}